+++ packages/adminui-vue/utils/lookups.js
@@ -28,6 +28,7 @@
  * @property {object} mapById
  * @property {string} descriptionAttrName
  */
+
 const Vue = require('vue')
 const UB = require('@unitybase/ub-pub')
 const ENUM_ENTITY = 'ubm_enum'
@@ -337,7 +338,7 @@ module.exports = {
    */
   getEnumItems (eGroup) {
     const items = instance.getMany(ENUM_ENTITY, { eGroup })
-    return _.orderBy(items, 'sortOrder').map(item => ({ code: item.code, name: item.name }))
+    return items.sort((a, b) => a.sortOrder - b.sortOrder).map(item => ({ code: item.code, name: item.name }))
   }
 }